+++ engine/consensus/sealing/core.go
@@ -29,12 +29,6 @@ import (
 	"github.com/onflow/flow-go/utils/logging"
 )
 
-// DefaultRequiredApprovalsForSealConstruction is the default number of approvals required to construct a candidate seal
-// for subsequent inclusion in block.
-// when set to 1, it requires at least 1 approval to build a seal
-// when set to 0, it can build seal without any approval
-// const DefaultRequiredApprovalsForSealConstruction = 1
-
 // DefaultEmergencySealingActive is a flag which indicates when emergency sealing is active, this is a temporary measure
 // to make fire fighting easier while seal & verification is under development.
 const DefaultEmergencySealingActive = false
